Updated gateway, spoke gateway and transit gateway

+++ aviatrix/resource_spoke_vpc.go
@@ -48,18 +48,27 @@ func resourceAviatrixSpokeVpc() *schema.Resource {
 			"subnet": {
 				Type:     schema.TypeString,
 				Required: true,
+				ForceNew: true,
 			},
 			"enable_nat": {
 				Type:     schema.TypeString,
 				Optional: true,
+				Computed: true,
 			},
 			"ha_subnet": {
 				Type:     schema.TypeString,
 				Optional: true,
+				Computed: true,
 			},
 			"ha_gw_size": {
 				Type:     schema.TypeString,
 				Optional: true,
+				Computed: true,
+			},
+			"single_az_ha": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
 			},
 			"dns_server": {
 				Type:     schema.TypeString,
@@ -68,11 +77,13 @@ func resourceAviatrixSpokeVpc() *schema.Resource {
 			"transit_gw": {
 				Type:     schema.TypeString,
 				Optional: true,
+				Computed: true,
 			},
 			"tag_list": {
 				Type:     schema.TypeList,
 				Elem:     &schema.Schema{Type: schema.TypeString},
 				Optional: true,
+				Computed: true,
 			},
 			"cloud_instance_id": {
 				Type:     schema.TypeString,
@@ -95,8 +106,12 @@ func resourceAviatrixSpokeVpcCreate(d *schema.ResourceData, meta interface{}) er
 		Subnet:         d.Get("subnet").(string),
 		HASubnet:       d.Get("ha_subnet").(string),
 		EnableNAT:      d.Get("enable_nat").(string),
+		SingleAzHa:     d.Get("single_az_ha").(string),
 		DnsServer:      d.Get("dns_server").(string),
 		TransitGateway: d.Get("transit_gw").(string),
+	}
+	if gateway.EnableNAT != "yes" {
+		gateway.EnableNAT = "no"
 	}
 	if cloudType := d.Get("cloud_type").(int); cloudType == 1 {
 		gateway.VnetRsrcGrp = ""
@@ -106,37 +121,40 @@ func resourceAviatrixSpokeVpcCreate(d *schema.ResourceData, meta interface{}) er
 		gateway.VpcID = ""
 		d.Set("vpc_id", gateway.VpcID)
 	}
-	if _, ok := d.GetOk("tag_list"); ok {
-		tagList := d.Get("tag_list").([]interface{})
-		tagListStr := goaviatrix.ExpandStringList(tagList)
-		gateway.TagList = strings.Join(tagListStr, ",")
-	}
 	log.Printf("[INFO] Creating Aviatrix Spoke VPC: %#v", gateway)
 
 	err := client.LaunchSpokeVpc(gateway)
 	if err != nil {
 		return fmt.Errorf("failed to create Aviatrix Spoke VPC: %s", err)
 	}
-	haSubnet := d.Get("ha_subnet").(string)
+	if enableNAT := d.Get("enable_nat").(string); enableNAT == "yes" {
+		log.Printf("[INFO] Aviatrix NAT enabled gateway: %#v", gateway)
+	}
+	if DNSServer := d.Get("dns_server").(string); DNSServer != "" {
+		log.Printf("[INFO] Aviatrix gateway DNS server: %#v", gateway)
+	}
 	haGwSize := d.Get("ha_gw_size").(string)
-	transitGwName := d.Get("transit_gw").(string)
-	d.SetId(gateway.GwName)
-	d.Set("ha_subnet", "")
-	d.Set("ha_gw_size", "")
-	d.Set("transit_gw", "")
-	if haSubnet != "" {
+	if singleAZHA := d.Get("single_az_ha").(string); singleAZHA == "enabled" {
+		singleAZGateway := &goaviatrix.Gateway{
+			GwName:   d.Get("gw_name").(string),
+			SingleAZ: d.Get("single_az_ha").(string),
+		}
+		log.Printf("[INFO] Enable Single AZ GW HA: %#v", singleAZGateway)
+		err := client.EnableSingleAZGateway(singleAZGateway)
+		if err != nil {
+			return fmt.Errorf("failed to create single AZ GW HA: %s", err)
+		}
+	}
+	if haSubnet := d.Get("ha_subnet").(string); haSubnet != "" {
 		//Enable HA
 		haGateway := &goaviatrix.SpokeVpc{
 			GwName:   d.Get("gw_name").(string),
 			HASubnet: haSubnet,
 		}
-
 		err = client.EnableHaSpokeVpc(haGateway)
 		if err != nil {
 			return fmt.Errorf("failed to enable HA Aviatrix TransitVpc: %s", err)
 		}
-		d.Set("ha_subnet", haSubnet)
-		d.Set("ha_gw_size", gateway.VpcSize)
 
 		log.Printf("[INFO]Resizing Spoke HA Gateway: %#v", haGwSize)
 		if haGwSize != gateway.VpcSize {
@@ -157,14 +175,30 @@ func resourceAviatrixSpokeVpcCreate(d *schema.ResourceData, meta interface{}) er
 			d.Set("ha_gw_size", haGwSize)
 		}
 	}
+	d.SetId(gateway.GwName)
 
-	if transitGwName != "" {
+	if _, ok := d.GetOk("tag_list"); ok {
+		tagList := d.Get("tag_list").([]interface{})
+		tagListStr := goaviatrix.ExpandStringList(tagList)
+		gateway.TagList = strings.Join(tagListStr, ",")
+		tags := &goaviatrix.Tags{
+			CloudType:    1,
+			ResourceType: "gw",
+			ResourceName: d.Get("gw_name").(string),
+			TagList:      gateway.TagList,
+		}
+		err = client.AddTags(tags)
+		if err != nil {
+			return fmt.Errorf("failed to add tags: %s", err)
+		}
+	}
+
+	if transitGwName := d.Get("transit_gw").(string); transitGwName != "" {
 		//No HA config, just return
 		err := client.SpokeJoinTransit(gateway)
 		if err != nil {
 			return fmt.Errorf("failed to join TransitVpc: %s", err)
 		}
-		d.Set("transit_gw", transitGwName)
 	}
 
 	return resourceAviatrixSpokeVpcRead(d, meta)
@@ -187,10 +221,49 @@ func resourceAviatrixSpokeVpcRead(d *schema.ResourceData, meta interface{}) erro
 	log.Printf("[TRACE] reading spoke gateway %s: %#v",
 		d.Get("gw_name").(string), gw)
 	if gw != nil {
+		d.Set("subnet", gw.VpcNet)
 		d.Set("vpc_size", gw.GwSize)
 		d.Set("public_ip", gw.PublicIP)
 		d.Set("cloud_instance_id", gw.CloudnGatewayInstID)
+		d.Set("enable_nat", gw.EnableNat)
+		d.Set("dns_server", gw.DnsServer)
+
+		if gw.SingleAZ == "yes" {
+			d.Set("single_az_ha", "enabled")
+		} else {
+			d.Set("single_az_ha", "disabled")
+		}
+		if gw.DnsServer != "" {
+			d.Set("dns_server", gw.DnsServer)
+		}
 	}
+
+	if gw.SpokeVpc == "yes" {
+		d.Set("transit_gw", gw.TransitGwName)
+	} else {
+		d.Set("transit_gw", "")
+	}
+
+	tags := &goaviatrix.Tags{
+		CloudType:    1,
+		ResourceType: "gw",
+		ResourceName: d.Get("gw_name").(string),
+	}
+	tagList, err := client.GetTags(tags)
+	if err != nil {
+		return fmt.Errorf("unable to read tag_list for gateway: %v due to %v", gateway.GwName, err)
+	}
+	var tagListStr []string
+	if _, ok := d.GetOk("tag_list"); ok {
+		tagList1 := d.Get("tag_list").([]interface{})
+		tagListStr = goaviatrix.ExpandStringList(tagList1)
+	}
+	if len(goaviatrix.Difference(tagListStr, tagList)) != 0 || len(goaviatrix.Difference(tagList, tagListStr)) != 0 {
+		d.Set("tag_list", tagList)
+	} else {
+		d.Set("tag_list", tagListStr)
+	}
+
 	haGateway := &goaviatrix.Gateway{
 		AccountName: d.Get("account_name").(string),
 		GwName:      d.Get("gw_name").(string) + "-hagw",
@@ -205,8 +278,8 @@ func resourceAviatrixSpokeVpcRead(d *schema.ResourceData, meta interface{}) erro
 		return fmt.Errorf("couldn't find Aviatrix SpokeVpc HA Gateway: %s", err)
 	}
 	log.Printf("[INFO] Spoke HA Gateway size: %s", haGw.GwSize)
+	d.Set("ha_subnet", haGw.VpcNet)
 	d.Set("ha_gw_size", haGw.GwSize)
-	//d.SetPartial("ha_subnet")
 
 	return nil
 }
@@ -224,6 +297,45 @@ func resourceAviatrixSpokeVpcUpdate(d *schema.ResourceData, meta interface{}) er
 	log.Printf("[INFO] Updating Aviatrix gateway: %#v", gateway)
 
 	d.Partial(true)
+	if d.HasChange("cloud_type") {
+		return fmt.Errorf("updating cloud_type is not allowed")
+	}
+	if d.HasChange("account_name") {
+		return fmt.Errorf("updating account_name is not allowed")
+	}
+	if d.HasChange("gw_name") {
+		return fmt.Errorf("updating gw_name is not allowed")
+	}
+	if d.HasChange("vpc_id") {
+		return fmt.Errorf("updating vpc_id is not allowed")
+	}
+	if d.HasChange("vpc_reg") {
+		return fmt.Errorf("updating vpc_reg is not allowed")
+	}
+	if d.HasChange("vnet_and_resource_group_names") {
+		return fmt.Errorf("updating vnet_and_resource_group_names is not allowed")
+	}
+	if d.HasChange("subnet") {
+		return fmt.Errorf("updating subnet is not allowed")
+	}
+	if d.HasChange("dns_server") {
+		return fmt.Errorf("updating dns_server is not allowed")
+	}
+
+	if d.HasChange("single_az_ha") {
+		if singleAz := d.Get("single_az_ha").(string); singleAz == "enabled" {
+			singleAZGateway := &goaviatrix.Gateway{
+				GwName:   d.Get("gw_name").(string),
+				SingleAZ: singleAz,
+			}
+			log.Printf("[INFO] Enable Single AZ GW HA: %#v", singleAZGateway)
+			err := client.EnableSingleAZGateway(singleAZGateway)
+			if err != nil {
+				return fmt.Errorf("failed to enable single AZ GW HA: %s", err)
+			}
+		}
+	}
+
 	if d.HasChange("tag_list") {
 		tags := &goaviatrix.Tags{
 			CloudType:    1,
@@ -239,17 +351,25 @@ func resourceAviatrixSpokeVpcUpdate(d *schema.ResourceData, meta interface{}) er
 		}
 		os := o.([]interface{})
 		ns := n.([]interface{})
-		oldTagList := goaviatrix.ExpandStringList(os)
-		tags.TagList = strings.Join(oldTagList, ",")
-		err := client.DeleteTags(tags)
-		if err != nil {
-			return fmt.Errorf("failed to delete tags : %s", err)
-		}
-		newTagList := goaviatrix.ExpandStringList(ns)
-		tags.TagList = strings.Join(newTagList, ",")
-		err = client.AddTags(tags)
-		if err != nil {
-			return fmt.Errorf("failed to add tags : %s", err)
+		oldList := goaviatrix.ExpandStringList(os)
+		newList := goaviatrix.ExpandStringList(ns)
+		oldTagList := goaviatrix.Difference(oldList, newList)
+		newTagList := goaviatrix.Difference(newList, oldList)
+		if len(oldTagList) != 0 || len(newTagList) != 0 {
+			if len(oldTagList) != 0 {
+				tags.TagList = strings.Join(oldTagList, ",")
+				err := client.DeleteTags(tags)
+				if err != nil {
+					return fmt.Errorf("failed to delete tags : %s", err)
+				}
+			}
+			if len(newTagList) != 0 {
+				tags.TagList = strings.Join(newTagList, ",")
+				err := client.AddTags(tags)
+				if err != nil {
+					return fmt.Errorf("failed to add tags : %s", err)
+				}
+			}
 		}
 		d.SetPartial("tag_list")
 	}
@@ -321,6 +441,26 @@ func resourceAviatrixSpokeVpcUpdate(d *schema.ResourceData, meta interface{}) er
 		}
 		d.SetPartial("ha_gw_size")
 	}
+	if d.HasChange("enable_nat") {
+		gw := &goaviatrix.Gateway{
+			CloudType: d.Get("cloud_type").(int),
+			GwName:    d.Get("gw_name").(string),
+		}
+		o, n := d.GetChange("enable_nat")
+		if o == "yes" && n == "no" {
+			err := client.DisableSNat(gw)
+			if err != nil {
+				return fmt.Errorf("failed to disable SNAT: %s", err)
+			}
+		}
+		if o == "no" && n == "yes" {
+			err := client.EnableSNat(gw)
+			if err != nil {
+				return fmt.Errorf("failed to enable SNAT: %s", err)
+			}
+		}
+		d.SetPartial("vpc_size")
+	}
 	if d.HasChange("transit_gw") {
 		spokeVPC := &goaviatrix.SpokeVpc{
 			CloudType:      d.Get("cloud_type").(int),
@@ -355,11 +495,11 @@ func resourceAviatrixSpokeVpcUpdate(d *schema.ResourceData, meta interface{}) er
 			}
 		}
 		d.SetPartial("transit_gw")
-
 	}
 	d.Partial(false)
-	//d.SetId(gateway.GwName)
-	return nil
+
+	d.SetId(gateway.GwName)
+	return resourceAviatrixSpokeVpcRead(d, meta)
 }
 
 func resourceAviatrixSpokeVpcDelete(d *schema.ResourceData, meta interface{}) error {
